Compilers update:
* Support x64
* Show all missing thm (KD87)
* Show all missing textures (KD87)
* xrSaveNodes:  reserve memory for compressed_nodes (abramcumner)
* setup_recalculationflags_file: use static buffer (abramcumner)
* xrPhase_UVmap: fix progress set (abramcumner)
* Change algorithm for calculating the area of a triangle (abramcumner)
* use nvtt for xrDXT
* MMX math replaced by SSE2
* Remove -keep_temp_files and add `-clear_temp_files`
* Support -f -s for xrAI
* Removed thread limit
* New lightmap save algorithm (DimaKuzmin)

+++ src/utils/xrLC_Light/xrLight_ImplicitCalcGlobs.cpp
@@ -51,7 +51,7 @@ void	ImplicitCalcGlobs::read				( INetReader	&r )
 		R_ASSERT( defl );
 		Fbox2 bounds;
 		defl->Bounds_Summary			(bounds);
-		Hash().initialize	(bounds,defl->faces.size());
+		Hash().initialize	(bounds,(u32)defl->faces.size());
 		for (u32 fid=0; fid<defl->faces.size(); fid++)
 		{
 			Face* F				= defl->faces[fid];
